+++ src/selectorValueChecker.ts
@@ -1,5 +1,6 @@
 import { Logging } from 'homebridge';
 import puppeteer from 'puppeteer-core';
+import { runStepsBeforeCheck } from './runStepsBeforeCheck';
 import { ChangeCheck } from './types/optionTypes';
 
 require('dotenv').config();
@@ -8,8 +9,9 @@ type getValueFromPageProps = {
     executablePath: string;
     changeCheck: ChangeCheck;
     waitForSelectorTimeout?: number;
-    log: Logging | Console['log'];
+    log: Logging;
     verboseLogging?: boolean;
+    previousValue?: string;
 };
 
 export const getValueFromPage = async (props: getValueFromPageProps) => {
@@ -25,6 +27,12 @@ export const getValueFromPage = async (props: getValueFromPageProps) => {
         const page = await browser.newPage();
         if (props.verboseLogging) { props.log('Browser initialized'); }
         await page.goto(props.changeCheck.url, { waitUntil: 'networkidle2' });
+
+        // Execute steps before actual check
+        if (props.changeCheck.stepsBeforeCheck) {
+            await runStepsBeforeCheck(page, props.changeCheck.stepsBeforeCheck);
+        }
+
         await page.waitForSelector(props.changeCheck.selector, {
             timeout: props.waitForSelectorTimeout || 30000 // 30 seconds
         });
@@ -34,11 +42,11 @@ export const getValueFromPage = async (props: getValueFromPageProps) => {
         const element = await page.$(props.changeCheck.selector);
         const text = await page.evaluate(element => { return element.textContent; }, element);
 
-        foundValue = text;
+        foundValue = text.replace(/(?:\r\n|\r|\n)/g, '').replace(/\s+/g, ' ').trim();
     } catch (e) {
-        props.log(e);
+        props.log.warn(e.toString());
 
-        foundValue = undefined;
+        foundValue = props.previousValue;
     } finally {
         await browser.close();