+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.5] - 2021-03-13
+### Added
+- `stepsBeforeCheck` option for every changeCheck. See README.md for how to use it.
+### Changed
+- Prevent motion detection when a timeout occurs
+
 ## [1.0.4] - 2021-03-11
 ### Added
 - Showing extra message when Chromium is not installed